Extract policy ladder execution into dedicated policy_ladder module.
Move solve loop and seed/band helpers out of analytic.py so turn assembly stays separate from tier policy execution.
